+++ src/unified-styles.css
@@ -8,6 +8,40 @@ body {
     letter-spacing: 0.3px;
 }
 
+/* 드라이브 상태 표시기 스타일 */
+.status-indicator {
+    display: flex;
+    align-items: center;
+    padding: 12px 16px;
+    border-radius: 8px;
+    margin-bottom: 8px;
+}
+
+.status-indicator.disconnected {
+    background-color: rgba(220, 53, 69, 0.1);
+    border: 1px solid rgba(220, 53, 69, 0.3);
+    color: #dc3545;
+}
+
+.status-indicator.loading {
+    background-color: rgba(13, 110, 253, 0.1);
+    border: 1px solid rgba(13, 110, 253, 0.3);
+    color: #0d6efd;
+}
+
+.status-indicator i {
+    margin-right: 8px;
+    font-size: 16px;
+}
+
+.drive-help {
+    margin-top: 8px;
+}
+
+.drive-help small {
+    color: rgba(255, 255, 255, 0.7);
+}
+
 .card {
     border-radius: 10px;
     border: none;
